Add Windows Compatible ID support
This allows to use Bricks on Windows 8 without installing an extra
driver. The compatible ID allows Windows 8 to autodetect a Brick as
WinUSB compatible and automatically install the correct driver.

Update the revision vom 0x100 to 0x110, because Windows caches the
(non-)existance of the 0xEE string descriptor based on PID+VID+Rev.