Circumvent bug in GCC 4.7.X
The uid_get_uid32 function has to be called from ram (i.e. it is in the section .ramfunc). To call a function in ram a long-call has to be executed, since the logical address distance from flash to ram is greater then 64M. This long-call is not done directly by GCC 4.7.X if the called function is small.

To circumvent this problem, we disable long-calls with "#pragma no_long_calls". In this case GCC generates an indirect call. The indirect call is slower, but it works...

As far as i can tell the Brick and Bricklet source code can now be compiled with all versions of GCC. Hooray!