Fix Bricklet plugin loading on SAM4S
Due to the global erase that is done on a SAM4S before loading
the Bricklet plugins into the flash, each flash page can only
be written once. In contrast, on SAM3S each page is individually
erased before writing, thus, each page can be written multiple
times.

This difference affects the loading of the last page of each
Bricklet plugin. Because in the last page the 3 struct pointers
for each Bricklet are also located. Those pointers were written
to the flash as a second step after loading the plugin. This
works on SAM3S but not on SAM4S, because each page can only be
written once.

Therefore the last page of the plugin has to be patched with the
pointers so it can be written in once.