Use DMA for synchronous I2C communication.
With the new implementation we can be interrupted by an interrupt
at any time without loosing data or reading/writing too much data.

This fixes a bug if the Master is RS485 slave, in this case it
was possible that one too much byte was read in some Bricklets
that use I2C. The result was a erroneous value.