Make usb_send_buffer_length variable volatile
Since this variable is changed in an interrupt and was not declared
volatile, the compiler did cache the variable locally in the
usb_handle_send function.

If the interrupt happend at the wrong time this could lead to duplicated
messages. A duplicated getter response will be discarded by the bindings,
but a duplicated callback can reach the user program!